engine/Object3D: implemented kind of LOD for skinning, e.g. skinning gets only computed every 2 frames if >= 10 meter away from camera, or every 4 frames if cam is from 20 meters far from camera, this is configurable now also

+++ src/tdme/engine/Timing.cpp
@@ -30,7 +30,7 @@ void Timing::updateTiming()
 	frame++;
 }
 
-int32_t Timing::getFrame()
+int64_t Timing::getFrame()
 {
 	return frame;
 }
